Add app configuration system
This adds a new config property to the Application class. Each subclass
must provide a static property defaultConfig that contains the default
configuration for that app. An app constructor should accept a config
object that passed config merged with the default config to the
superclass via super(Object.assign(AppClass.defaultConfig, config)).