Fix overwriting static defaultConfig in subclasses
In ES6, in contrary to many other programming languages, static class
members are actually inherited by subclasses. Accessing
SubClass.defaultConfig actually points to BaseClass.defaultConfig such
that each app modified the defaultConfig of each base class.

The solution is to rely on a get-method. These are also inherited, but
at least can be properly overwriten in a subclass.