Fix various issues with car-park
Previously, a new random level was chosen whenever a new difficulty
was selected. This is very confusing if you move the slider up and down,
because new levels appear all the time. It prevents the user from
repeating a level, even if the slider was just clicked accidentally.
Colors also changed all the time, so even if the same level appeared
twice, it would be hard to identify.

The new mechanism will choose the colors and levels only during reset.

By introducing two global variables numlevels and numdifficulties, it is
now possible to change the number of levels and difficulties
independently. To match both numbers, several levels of the same
difficulty might be chosen. There are currently 7 levels chosen out of
5 difficilties.

I tried to keep the changes as small as possible. However, it was
hardly possible to make any modification without refactoring the code.
I ran into a couple of bugs, which were hard to fix in isolation.
E.g. it was possible to move the slider knob above and below the slider,
which triggered out-of-bound errors in certain situations.
Implementation details of the sliders were spread over several places,
which made it hard to fix it without doing a proper refactoring.
I also considerably reduced the number of global variables. They are now
all listed and initialized at the top of the init script.