Move level selection from CindyScript to carparklevels plugin
The previous plugin function 'randomlevel' is replaced with a
'getshuffledlevels' function which will be called during reset of the
app. The app will adapt itself with respect to the number of levels
returned such that the number of levels can now be adjusted during
runtime. A new config parameter 'numLevels' is introcuded to be able to
adjust the number of levels independently for each app instance.

The 'shuffledlevels' function will ensure, in contrast to 'randomlevel',
that no level is included twice unless there are no more levels to
choose from. It uses a modified Fisher-Yates algorithm for efficiency.