Refactor sphere-chaos
Interactions with the geometric construction are now done directly
via Cinderella instead of CindyScript. This removes a lot of complexity from the draw, mousedown and mouseup scripts.

The code can now deal with arbitrary numbers of reflective
circles.

Many functions have been cleaned up and a lot of global variables
have been removed in the process.