Add system to retrieve config overrides
The Application class now offers a static method getConfigOverrides()
for retrieving config override data. Each subclass that wants to make
use of this functionality needs to implement the function
`static async retrieveConfigOverrides()` that will be called when
`ApplicationSubClass.getConfigOverrides()` is called (static methods are
inherited in JavaScript such that subclasses do not need to implement
this function themselves).

Most apps will host their config override data next to the class file.
Therefore, subclasses of CindyApp can use the following implementation
of `retrieveConfigOverrides`:
```
static async retrieveConfigOverrides() {
    return this.retrieveConfigOverridesJsonByClass(import.meta.url);
}
```
It loads the file `AppClassOverrides.json`, where `AppClass` is replaced
by the actual name of the class this function is called on. So `MyApp`
would load a file called `MyAppOverrides.json`.