Fix CindyApp loading before the canvas became a descendant of document.body
The previous observer only checked for the canvas having a parent.
However, this is already true when the canvas becomes a child of another
element, but not necessarily a descendant of document.body. This is now
properly checked by the oberserver callback using
document.body.contains(canvas). This finally ensures that
canvas.clientWidth and canvas.clientHeight are non-zero when the canvas
is passed to CindyJS.