Fix out-of-range behaviour of level selector
Previously, the discrete position of the slider knob has been computed
via the distance to the lower point of the slider line segment. When
moving below the lower point, this distance increased again such that
the level index increased as well instead of being stuck at 1.

The new algorithm is a lot more stable and can easily be reused
elsewhere: It computes the orthogonal projection of the slider knob
position onto the slider segment and determines the linear
interpolation parameter of this projection with respect to the slider
line segment (0 at the beginning, 1 at the end). The interpolation
parameter is then properly rounded to compute a discrete slot value.

The computation is independent of the slider direction. Horizontal,
vertical and diagonal sliders all work.